pyproject.toml
[project]
name = "cutedsl-gpt"
version = "0.1.0"
description = "Learning NVIDIA CuTe DSL by writing a GPT from scratch (Blackwell examples)"
readme = "README.md"
requires-python = ">=3.10,<3.13"
authors = [{ name = "Anton Krasilnikov", email = "akrasilnikovvv@gmail.com" }]

dependencies = [
    # CuTe DSL (Python). Blackwell kernels require a recent release.
    "nvidia-cutlass-dsl>=4.1.0",

    # Core numerics / tensors
    "torch>=2.5.0",
    "numpy>=1.26",

    # GPT training utilities
    "tiktoken>=0.7.0",
    "datasets>=2.20",
    "tqdm>=4.66",
    "einops>=0.8",

    # Experiment tracking & config
    "wandb>=0.17",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-xdist>=3.5",
    "ruff>=0.6",
    "mypy>=1.10",
    "ipykernel>=6.29",
    "jupyterlab>=4.2",
    "matplotlib>=3.9",
]

[project.urls]
"CuTe DSL Examples" = "https://github.com/NVIDIA/cutlass/tree/main/examples/python/CuTeDSL/blackwell"
"CUTLASS" = "https://github.com/NVIDIA/cutlass"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cutedsl_gpt"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true
